Move fields file out of fields folder
