+++ packages/data/src/types/abstract-query/fields.ts
@@ -1,0 +1,9 @@
+import type { AbstractQueryFieldNodeFn } from './fields/function.js';
+import type { AbstractQueryFieldNodeNestedMany, AbstractQueryFieldNodeNestedOne } from './fields/nested.js';
+import type { AbstractQueryFieldNodePrimitive } from './fields/primitive.js';
+
+export type AbstractQueryFieldNode =
+	| AbstractQueryFieldNodePrimitive
+	| AbstractQueryFieldNodeFn
+	| AbstractQueryFieldNodeNestedMany
+	| AbstractQueryFieldNodeNestedOne;
