--- packages/data/src/types/abstract-query/fields/fields.ts
@@ -1,9 +1,0 @@
-import type { AbstractQueryFieldNodeFn } from './function.js';
-import type { AbstractQueryFieldNodeNestedMany, AbstractQueryFieldNodeNestedOne } from './nested.js';
-import type { AbstractQueryFieldNodePrimitive } from './primitive.js';
-
-export type AbstractQueryFieldNode =
-	| AbstractQueryFieldNodePrimitive
-	| AbstractQueryFieldNodeFn
-	| AbstractQueryFieldNodeNestedMany
-	| AbstractQueryFieldNodeNestedOne;
