+++ packages/data/src/types/abstract-query/index.ts
@@ -1,3 +1,4 @@
+export * from './fields.js';
 export * from './fields/index.js';
 export * from './modifiers.js';
 export * from './modifiers/index.js';
